+++ react/src/pages/auth/index.js
@@ -1,2 +1,2 @@
-export * from "@/pages/auth/sign-in";
-export * from "@/pages/auth/sign-up";
+export * from "@/pages/auth/login";
+export * from "@/pages/auth/register";
